More async refactoring
